Optimise UM2+ polypropylene profile
Settings that are common to all UM2+ polypropylene profiles are put here. They also happen to be exactly equal for all different nozzle sizes so the settings can be put in the Ultimaker 2+ section. And in fact, a lot of those settings are the same as Ultimaker 3 so I've put those settings at the very top to be shared with all printers. This means that even third-party printers get sort-of correct material settings for polypropylene.

Contributes to issue CURA-3960.