Remove warnings about generic profiles from non-generic profiles
These are not generic. Some of these had a warning that it's auto-generated (read: copy-pasted) and may therefore be incorrect, but we're pretty sure that these are sort of correct and that every profile ever may be incorrect, so I deem those superfluous and removed them as well.

Contributes to issue CURA-3960.