Add descriptions and adhesion info for PLA profiles
These are taken from an internal document about the different materials, which took it from the Ultimaker website description. I applied it to all PLA profiles, even third-party profiles, since it should be more or less the same. The only exceptions I made are for Polywood and Polyflex since those are woodfill and flexible variants and could have very different properties.

Contributes to issue CURA-3960.