Fix indenting
Some tabs mixed with spaces here and there, and even worse, using a mix of 1, 2, 3 and 4 spaces for indenting. This is why I generally prefer tabs, you just indent with one tab and the IDE does the rest. But our code style specifies 4 spaces so that's what it's got to be.

Contributes to issue CURA-3960.