Add generic filled PLA
This material profile behaves a lot like PLA, with some subtle differences. Printers can make their own profiles for this and material manufacturers can base their profile on this one too. Its settings are the same as PLA, but with a different GUID, density and colour.

Contributes to issue CURA-6329.