Change shrinkage percentage overrides to Cura-specific settings
It turns out that the Ultimaker firmware was not actually using this setting, but it was validating it to be between 0 and 100%, basically validating that a material shouldn't shrink to less than 50% of its original size. When we changed the unit of this setting, this caused their firmware to no longer read these materials.
This change makes this use the Cura namespace and the Cura setting name, so the firmware doesn't recognise the setting any more.

Contributes to issue CURA-7724.