Retire old check_material_profiles script, Jenkinsfile
Girlscouting: retire the old version of the `check_material_profiles.py`
script, as it's not actually used anywhere. We currently depend on the
"with lxml" version of that script, so let that new version take its
place.

This also removes the old `Jenkinsfile`, which is no longer relevant to
our current CI setup.

Finally, remove comments that refer to the "lxml version not currently
being used in CI". This is just confusing.