[CES-3334] Update the docker and requirements to Bookworm
With this PR we update the docker file so that it builds on Debian
Bookworm instead of Debian Buster.

Buster is no longer provided or supported by Debian.

We also update the requirements to pull in a version of lxml that is
available for Debian Bookworm.

The Docker file now uses a python virtual environment to manage custom
package installs which is required by PEP-668.